amélioration routes post & comment

+++ backend/controllers/userController.js
@@ -103,15 +103,30 @@ exports.modifyUser = (req, res) => {
 };
 
 exports.deleteUser = (req, res) => {
-  User.destroy({ where: { uid: req.params.uid } })
+  User.findOne({ where: { uid: req.params.uid } })
     .then((user) => {
-      if (user) {
-        res.status(201).json({ message: "Utilisateur supprimé !" });
-      } else {
-        res.status(404).json({ message: "Erreur, utilisateur non trouvé !" });
-      }
+      bcrypt
+        .compare(req.body.password, user.password)
+        .then((valid) => {
+          if (!valid) {
+            return res.status(401).json({ error: "Mot de passe incorrect !" });
+          } else {
+            User.destroy({ where: { uid: req.params.uid } })
+              .then((user) => {
+                if (user) {
+                  res.status(201).json({ message: "Utilisateur supprimé !" });
+                } else {
+                  res
+                    .status(404)
+                    .json({ message: "Erreur, utilisateur non trouvé !" });
+                }
+              })
+              .catch((error) => res.status(400).json({ error }));
+          }
+        })
+        .catch((error) => res.status(500).json({ error }));
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getAllUsers = (req, res) => {
